Revert "[Feature] Added DetailScreen feature, and refactored App file by moving code to Screen and HomeScreen files"
