--- composeApp/src/commonMain/composeResources/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">KmpMovies</string>
-    <string name="back">Go back</string>
-</resources>
